NJ: swap old scraper for new
